requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
openai>=1.0.0
cohere>=4.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
requests>=2.28.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
numpy>=1.26.0